[svn-r304] ./test/hyperslab.c 	Added UL->L casts in six places to quiet warnings on Irix -64.

+++ test/hyperslab.c
@@ -790,8 +790,8 @@ test_transpose(size_t nx, size_t ny)
     size[1] = ny;
     src_stride[0] = 0;
     src_stride[1] = sizeof(*src);
-    dst_stride[0] = (1 - nx * ny) * sizeof(*src);
-    dst_stride[1] = nx * sizeof(*src);
+    dst_stride[0] = (ssize_t)((1 - nx * ny) * sizeof(*src));
+    dst_stride[1] = (ssize_t)(nx * sizeof(*src));
 
     /* Copy and transpose */
     if (nx == ny) {
@@ -891,7 +891,7 @@ test_sub_super(size_t nx, size_t ny)
     /* Setup */
     size[0] = nx;
     size[1] = ny;
-    src_stride[0] = 2 * ny;
+    src_stride[0] = (ssize_t)(2 * ny);
     src_stride[1] = 2;
     dst_stride[0] = 0;
     dst_stride[1] = 1;
@@ -939,9 +939,9 @@ test_sub_super(size_t nx, size_t ny)
     src_stride[1] = 1;
     src_stride[2] = 0;
     src_stride[3] = 0;
-    dst_stride[0] = 2 * ny;
-    dst_stride[1] = 2 * sizeof(uint8) - 4 * ny;
-    dst_stride[2] = 2 * ny - 2 * sizeof(uint8);
+    dst_stride[0] = (ssize_t)(2 * ny);
+    dst_stride[1] = (ssize_t)(2 * sizeof(uint8) - 4 * ny);
+    dst_stride[2] = (ssize_t)(2 * ny - 2 * sizeof(uint8));
     dst_stride[3] = sizeof(uint8);
 
     /* Copy */
